+++ src/services/Redirects.php
@@ -11,6 +11,7 @@
 
 namespace nystudio107\retour\services;
 
+use craft\base\ElementInterface;
 use nystudio107\retour\Retour;
 use nystudio107\retour\events\RedirectEvent;
 use nystudio107\retour\events\ResolveRedirectEvent;
@@ -150,6 +151,16 @@ class Redirects extends Component
      * @var null|array
      */
     protected $cachedStaticRedirects;
+
+    /**
+     * @var null|array
+     */
+    protected $cachedRegExRedirects;
+
+    /**
+     * @var null|array
+     */
+    protected $cachedExactMatchRedirects;
 
     // Public Methods
     // =========================================================================
@@ -342,8 +353,17 @@ class Redirects extends Component
 
             return $redirect;
         }
+
+        $redirect = $this->getStaticRedirect($fullUrl, $pathOnly, $siteId);
+        if ($redirect) {
+            $this->incrementRedirectHitCount($redirect);
+            $this->saveRedirectToCache($pathOnly, $redirect);
+
+            return $redirect;
+        }
+
         // Resolve static redirects
-        $redirects = $this->getAllStaticRedirects(null, $siteId);
+        $redirects = $this->getAllRegExRedirects(null, $siteId);
         $redirect = $this->resolveRedirect($fullUrl, $pathOnly, $redirects, $siteId);
         if ($redirect) {
             return $redirect;
@@ -624,6 +644,114 @@ class Redirects extends Component
     }
 
     /**
+     * @param string $fullUrl
+     * @param string $pathOnly
+     * @param $siteId
+     * @return mixed|null
+     */
+    public function getStaticRedirect(string $fullUrl, string $pathOnly, $siteId)
+    {
+        $staticCondition = ['redirectMatchType' => 'exactmatch'];
+        $siteCondition = [
+            'or',
+            ['siteId' => $siteId],
+            ['siteId' => null]
+        ];
+        $pathCondition = [
+            'or',
+            ['and',
+                ['redirectSrcMatch' => 'pathonly'],
+                ['redirectSrcUrlParsed' => $pathOnly]
+            ],
+            ['and',
+                ['redirectSrcMatch' => 'fullurl'],
+                ['redirectSrcUrlParsed' => $fullUrl]
+            ],
+        ];
+
+        $query = (new Query)
+            ->from('{{%retour_static_redirects}}')
+            ->where(['and',
+                $staticCondition,
+                $pathCondition,
+                $siteCondition
+            ])
+            ->limit(1);
+
+        return $query->one();
+    }
+
+    /**
+     * @param null|int $limit
+     * @param int|null $siteId
+     *
+     * @return array All of the regex match redirects
+     */
+    public function getAllRegExRedirects(int $limit = null, int $siteId = null): array
+    {
+        // Cache it in our class; no need to fetch it more than once
+        if ($this->cachedRegExRedirects !== null) {
+            return $this->cachedRegExRedirects;
+        }
+
+        $redirects = $this->getRedirectsByMatchType($limit, $siteId, 'regexmatch');
+
+        // Cache for future accesses
+        $this->cachedRegExRedirects = $redirects;
+
+        return $redirects;
+    }
+
+    /**
+     * @param null|int $limit
+     * @param int|null $siteId
+     *
+     * @return array All of the regex match redirects
+     */
+    public function getAllExactMatchRedirects(int $limit = null, int $siteId = null): array
+    {
+        // Cache it in our class; no need to fetch it more than once
+        if ($this->cachedExactMatchRedirects !== null) {
+            return $this->cachedExactMatchRedirects;
+        }
+
+        $redirects = $this->getRedirectsByMatchType($limit, $siteId, 'exactmatch');
+
+        // Cache for future accesses
+        $this->cachedExactMatchRedirects = $redirects;
+
+        return $redirects;
+    }
+
+    /**
+     * @param int|null $limit
+     * @param int|null $siteId
+     * @param string $type
+     * @return array
+     */
+    protected function getRedirectsByMatchType(int $limit = null, int $siteId = null, string $type): array
+    {
+        // Query the db table
+        $query = (new Query())
+            ->from(['{{%retour_static_redirects}}'])
+            ->orderBy('redirectMatchType ASC, redirectSrcMatch ASC, hitCount DESC');
+
+        if ($siteId) {
+            $query
+                ->where(['siteId' => $siteId])
+                ->orWhere(['siteId' => null]);
+        }
+
+        if ($limit) {
+            $query->limit($limit);
+        }
+
+        $query->andWhere(['redirectMatchType' => $type]);
+
+        return $query->all();
+    }
+
+    /**
      * @param null|int $limit
      * @param int|null $siteId
      *
@@ -686,7 +814,7 @@ class Redirects extends Component
         $query = (new Query())
             ->from(['{{%retour_static_redirects}}'])
             ->where(['redirectSrcUrl' => $redirectSrcUrl])
-            ;
+        ;
         if ($siteId) {
             $query
                 ->andWhere(['or', [
@@ -698,6 +826,28 @@ class Redirects extends Component
         $redirect = $query->one();
 
         return $redirect;
+    }
+
+    /**
+     * Return redirects for a given element.
+     *
+     * @param int $elementId
+     * @param int|null $siteId
+     *
+     * @return null|array
+     */
+    public function getRedirectsByElementId(int $elementId, int $siteId = null)
+    {
+        // Query the db table
+        $query = (new Query())
+            ->from(['{{%retour_static_redirects}}'])
+            ->where(['associatedElementId' => $elementId])
+        ;
+        if ($siteId !== null) {
+            $query->andWhere(['siteId' => $siteId]);
+        }
+
+        return $query->all();
     }
 
     /**
@@ -760,6 +910,44 @@ class Redirects extends Component
                 Craft::debug('Rows affected: '.$rowsAffected, __METHOD__);
             } catch (\Exception $e) {
                 Craft::error($e->getMessage(), __METHOD__);
+            }
+        }
+    }
+
+    /**
+     * Save an element redirect.
+     *
+     * @param ElementInterface $element
+     * @param array $fieldValue
+     */
+    public function enableElementRedirect(ElementInterface $element, array $fieldValue)
+    {
+        $redirectConfig = [
+            'redirectMatchType' => 'exactmatch',
+            'redirectSrcUrl' => $fieldValue['redirectSrcUrl'],
+            'siteId' => $element->siteId,
+            'associatedElementId' => $element->getCanonicalId(),
+            'enabled' => true,
+            'redirectSrcMatch' => 'pathonly',
+            'redirectDestUrl' => $element->getUrl(),
+            'redirectHttpCode' => $fieldValue['redirectHttpCode'],
+        ];
+
+        $this->saveRedirect($redirectConfig);
+    }
+
+    /**
+     * Delete an element redirect.
+     *
+     * @param ElementInterface $element
+     */
+    public function removeElementRedirect(ElementInterface $element, bool $allSites = false)
+    {
+        $redirects = $this->getRedirectsByElementId($element->getCanonicalId(), $allSites ? null : $element->siteId);
+
+        if (!empty($redirects)) {
+            foreach ($redirects as $redirect) {
+                $this->deleteRedirectById($redirect['id']);
             }
         }
     }